clk: sunxi-ng: a64: try to keep TCON0 clock in CCU
As part of the clock tree of LCD and HDMI pipelines on A64 are
conflicted, usual way to keep LCD clock cannot work on A64.

However, there's a PLL dedicated to TCON0, PLL-MIPI, which can be
utilized to get a clock tree that can satisfy both clock rates on LCD
and HDMI pipelines.

This patch forbids the PLL-Video0-2X parent of TCON0 clock, and let TCON0
clock (and then PLL-MIPI by CLK_SET_RATE_PARENT) gets resetted when
PLL-Video0 changes (because of HDMI PHY clk which is a child of
PLL-Video0 and has CLK_SET_RATE_PARENT set). In this way we can get
clock tree to satisfy both pipelines.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>