gfs2: Work around ->readahead deadlock
With the conversion from ->readpages to ->readahead in commit d4388340ae0b
("fs: convert mpage_readpages to mpage_readahead"), gfs2_readahead is called
with the pages to read ahead locked.  This means that grabbing the inode glock
in gfs2_readahead can now deadlock with other processes which are holding the
inode glock and trying to lock pages.

The proper fix would be to move the inode glock into the ->read_iter file and
->fault vm operations.  However, gfs2 allows reads to be served out of the page
cache without taking the corresponding inode glock if the requested data is
cached, and there currently is no reasonable way to implement that conversion
without losing this important optimization.

So our only other option to avoid deadlocks is to use a trylock operation in
gfs2_readahead.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>