power: supply: ip5xxx: Handle powerdown state of the chip
When IP5209 goes to sleep, it will become unresponsive on I2C bus.
When used with non-I2C regmap, we are notified about this particular
state via -EXDEV error. This is not a fatal error state, so treat it
as if the battery is not charging, and boost is disabled, which
is what happens in reality.

Signed-off-by: Ondrej Jirman <megous@megous.com>