phy: phy-rockchip-inno-usb2: Decrease delay between port init and charger detection
When used on systems with TCPM, the charger detection via USB2.0 PHY
will take too much time during boot, so the fusb302 driver will not
be able to get the result, and will set incorrect current_max on
its power supply device.

This will lead to failure to charge when the device is powered up
while connected to non-PD charger.

Decreasing delay of otg_sm_work from 6s to 500ms ensures that
port type detection via the PHY will finish well in time for
TCPM to be able to use its result when Rp=default is detected
via CC pins.

Signed-off-by: Ondrej Jirman <megous@megous.com>